Moved ContractAddress to types
