+++ concordium-sdk/src/main/java/com/concordium/sdk/exceptions/ContractInstanceNotFoundException.java
@@ -1,8 +1,7 @@
 package com.concordium.sdk.exceptions;
 
-import com.concordium.sdk.responses.transactionstatus.ContractAddress;
+import com.concordium.sdk.types.ContractAddress;
 import com.concordium.sdk.transactions.Hash;
-import lombok.Getter;
 
 /**
  * A checked exception thrown when a contract instance could not be found with the {@link ContractAddress}
