+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/InterruptedResult.java
@@ -1,5 +1,7 @@
 package com.concordium.sdk.responses.transactionstatus;
 
+import com.concordium.sdk.types.AbstractAddress;
+import com.concordium.sdk.types.ContractAddress;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
