+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/RejectReasonInvalidContractAddress.java
@@ -1,10 +1,9 @@
 package com.concordium.sdk.responses.transactionstatus;
 
+import com.concordium.sdk.types.ContractAddress;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
-
-import java.util.Map;
 
 /**
  * Contract instance does not exist on chain.
