+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/TransferredResult.java
@@ -1,6 +1,7 @@
 package com.concordium.sdk.responses.transactionstatus;
 
 import com.concordium.sdk.transactions.CCDAmount;
+import com.concordium.sdk.types.AbstractAddress;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
