+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/Parameter.java
@@ -10,7 +10,10 @@ import java.nio.ByteBuffer;
 
 
 /**
+ * The parameters are used for updating the smart contract instance.
+ * i.e. calling a recieve function exposed in the smart contract with the parameters.
  * Buffer of the parameters message.
+ * In the current supported protocols the size is limited to be 1kb.
  */
 
 @Getter
