+++ concordium-sdk/src/main/java/com/concordium/sdk/types/ContractAddress.java
@@ -1,4 +1,4 @@
-package com.concordium.sdk.responses.transactionstatus;
+package com.concordium.sdk.types;
 
 import com.concordium.sdk.serializing.JsonMapper;
 import com.concordium.sdk.transactions.AccountType;
@@ -13,13 +13,14 @@ import lombok.ToString;
 import lombok.val;
 
 
+import java.nio.ByteBuffer;
 import java.util.Objects;
 import java.util.Optional;
 
 @Getter
 @ToString
 @EqualsAndHashCode
-public class ContractAddress extends AbstractAddress {
+public final class ContractAddress extends AbstractAddress {
 
     @JsonProperty("subindex")
     private final int subIndex;
@@ -55,5 +56,16 @@ public class ContractAddress extends AbstractAddress {
         } catch (JsonProcessingException e) {
             throw new IllegalArgumentException("Cannot parse ContractAddress Array JSON", e);
         }
+    }
+
+    public static ContractAddress from(int index, int subIndex) {
+        return new ContractAddress(index, subIndex);
+    }
+
+    public byte[] getBytes() {
+        val buffer = ByteBuffer.allocate(Long.BYTES + Long.BYTES );
+        buffer.putLong(this.getIndex());
+        buffer.putLong(this.getSubIndex());
+        return buffer.array();
     }
 }
