+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/UpdateContractPayloadTest.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.transactions;
 
+import com.concordium.sdk.types.ContractAddress;
 import lombok.val;
 import org.junit.Test;
