+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/UpdateContractTransactionTest.java
@@ -3,6 +3,7 @@ package com.concordium.sdk.transactions;
 import com.concordium.sdk.crypto.ed25519.ED25519ResultCode;
 import com.concordium.sdk.exceptions.ED25519Exception;
 import com.concordium.sdk.exceptions.TransactionCreationException;
+import com.concordium.sdk.types.ContractAddress;
 import com.concordium.sdk.types.UInt64;
 import lombok.SneakyThrows;
 import org.junit.Test;
